+++ commands/scopecommand.go
@@ -1,0 +1,5 @@
+package commands
+
+type ScopeCommand interface {
+	ParseCommand() string
+}
