--- interfaces/scopecommand.go
@@ -1,5 +1,0 @@
-package interfaces
-
-type ScopeCommand interface {
-	Command() string
-}
